feat: extract SurveyForm component from SurveyPage
